+++ spec/support_table_data_spec.rb
@@ -188,7 +188,7 @@ describe SupportTableData do
 
   describe "support_table_classes" do
     it "gets a list of all loaded support table classes with belongs to dependencies listed first" do
-      expect(SupportTableData.support_table_classes).to eq [Group, Hue, Color, Invalid]
+      expect(SupportTableData.support_table_classes).to eq [Group, Hue, Shade, Color, Invalid]
     end
   end