+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 1.3.1
+
+### Added
+
+- Added support for autosave associations. Data in autosave associations will be persisted when the support table is synced if it was changed by the support table data.
+
 ## 1.3.0
 
 ### Added